♻️ MAde own component
